+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Remove deprecated `Phake::partMock()`
 - Remove deprecated `AnswerProxyInterface::thenGetReturnByLambda()`
 - `Phake::verifyNoOtherInteractions()` is now variadic
+- Private methods will not be mocked anymore (there is no point to it).